moving code from infra_manager to api_integration
